refactor(cache): rename plugin folder cachePlugin -> cache
Match the bare-noun convention of the other plugin folders (access, email,
storage, usage, …). Folder only — the exported plugin function stays
`cachePlugin` (mirrors accessPlugin/storagePlugin). All `@/plugins/cachePlugin`
imports and doc references repointed to `@/plugins/cache`.

Refs #555
