+++ src/collections/Meditations/endpoints/lectures.ts
@@ -18,7 +18,7 @@ import type {
   SubtleSystemNode,
   UserChoice,
 } from '@/payload-types'
-import { CUSTOM_READS, publicReadCacheHeaders } from '@/plugins/cachePlugin'
+import { CUSTOM_READS, publicReadCacheHeaders } from '@/plugins/cache'
 import { asTrustedReq } from '@/plugins/usage/hooks'
 
 /**
