+++ src/payload.config.ts
@@ -17,7 +17,7 @@ import { createWorkerSafeLogger } from '@/lib/logger/workerSafeLogger'
 import { SUPPORTED_TIMEZONES } from '@/lib/timezones'
 import { getServerUrl } from '@/lib/utilities/serverUrl'
 import { accessPlugin, bypassPermissions, filterAvailableLocales } from '@/plugins/access'
-import { cachePlugin } from '@/plugins/cachePlugin'
+import { cachePlugin } from '@/plugins/cache'
 import { resendAdapter } from '@/plugins/email'
 import { openapiEndpointAuth, scalarPlugin } from '@/plugins/openapi'
 import { sentryPlugin } from '@/plugins/sentry'
